blorgo/pyproject.toml
[project]
name = "blorgo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "resonate-sdk>=0.5.4",
    "torch>=2.7.1",
    "transformers>=4.53.2",
]
